--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -103,13 +103,20 @@ jobs:
         benchstat main=bench-main.txt current=bench-current.txt > result.txt
         cat result.txt
 
+    - name: Run OpenAI
+      if: env.prExists == 'true'
+      run: |
+        echo "Running OpenAI"
+        benchresult=$(cat result.txt | jq -Rs .)
+        opencomment = $(curl -X POST 'https://webpoc.ashystone-186beaf0.eastus.azurecontainerapps.io/Bench' -H 'accept: text/plain' -H 'Content-Type: application/json' -d "{\"benchresult\": \"$benchresult\"}")
+  
     - name: Create PR Comment
       if: env.prExists == 'true'
       env:
         GITHUB_TOKEN: ${{ secrets.PERSONAL_ACCESS_TOKEN }}
       run: |
         if [ -s result.txt ]; then
-          COMMENT_BODY=$(echo -e "## Resultados do Benchmark\n\`\`\`" && cat result.txt && echo -e "\`\`\`")
+          COMMENT_BODY=$(echo -e "## Resultados do Benchmark\n\`\`\`" && cat result.txt && echo -e "\`\`\`\n\n## Análise do resultado \n\`\`\`$opencomment\`\`\`")
           echo "$COMMENT_BODY" | gh pr comment $PR_NUMBER --body-file=-
         else
           echo "result.txt is empty. Skipping comment creation."