--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -108,7 +108,7 @@ jobs:
       run: |
         echo "Running OpenAI"
         benchresult=$(cat result.txt | jq -Rs .)
-        opencomment = $(curl -X POST 'https://webpoc.ashystone-186beaf0.eastus.azurecontainerapps.io/Bench' -H 'accept: text/plain' -H 'Content-Type: application/json' -d "{\"benchresult\": \"$benchresult\"}")
+        opencomment=$(curl -X POST 'https://webpoc.ashystone-186beaf0.eastus.azurecontainerapps.io/Bench' -H 'accept: text/plain' -H 'Content-Type: application/json' -d "{\"benchresult\": \"$benchresult\"}")
   
     - name: Create PR Comment
       if: env.prExists == 'true'