--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -118,8 +118,9 @@ jobs:
       run: |
         if [ -s result.txt -a -s opencomment.md ]; then
           COMMENT_BODY=$(echo -e "## Resultados do Benchmark\n\n\`\`\`\n$(cat result.txt)\n\`\`\`\n\n## Análise do resultado\n\n\`\`\`\n$(cat opencomment.md)\n\`\`\`")
-          echo "Comment Body: $COMMENT_BODY"
-          echo "$COMMENT_BODY" | gh pr comment $PR_NUMBER --body -
+          echo "$COMMENT_BODY" > comment_body.md
+          cat comment_body.md
+          gh pr comment $PR_NUMBER --body-file comment_body.md
         else
           echo "result.txt or opencomment.md is empty. Skipping comment creation."
         fi