--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -117,7 +117,7 @@ jobs:
         GITHUB_TOKEN: ${{ secrets.PERSONAL_ACCESS_TOKEN }}
       run: |
         if [ -s result.txt -a -s opencomment.md ]; then
-          COMMENT_BODY=$(echo -e "## Resultados do Benchmark\n\n\`\`\`\n$(cat result.txt)\n\`\`\`\n\n## Análise do resultado\n\n\`\`\`\n$(cat opencomment.md)\n\`\`\`")
+          COMMENT_BODY=$(echo -e "## Resultados do Benchmark\n\n$(cat result.txt)\n\n## Análise do resultado\n\n$(cat opencomment.md)")
           echo "$COMMENT_BODY" > comment_body.md
           cat comment_body.md
           gh pr comment $PR_NUMBER --body "$(cat comment_body.md)"