--- a/app/main.go
+++ b/app/main.go
@@ -2,18 +2,44 @@ package main
 
 import "fmt"
 
-// bubbleSort ordena um slice de inteiros
+// Função para ordenar um array de inteiros
 func Sort(arr []int) []int {
-	n := len(arr)
-	// Percorre o slice comparando o elemento ao lado e trocando de lugar se necessário
-	for i := 0; i < n-1; i++ {
-		for currentIndex := 0; currentIndex < n-i-1; currentIndex++ {
-			if arr[currentIndex] > arr[currentIndex+1] {
-				arr[currentIndex], arr[currentIndex+1] = arr[currentIndex+1], arr[currentIndex]
-			}
+	if len(arr) <= 1 {
+		return arr
+	}
+
+	middle := len(arr) / 2
+	left := Sort(arr[:middle])
+	right := Sort(arr[middle:])
+
+	return merge(left, right)
+}
+
+// Função para mesclar dois arrays
+func merge(left, right []int) []int {
+	result := []int{}
+	i, j := 0, 0
+
+	for i < len(left) && j < len(right) {
+		if left[i] < right[j] {
+			result = append(result, left[i])
+			i++
+		} else {
+			result = append(result, right[j])
+			j++
 		}
 	}
-	return arr
+
+	for i < len(left) {
+		result = append(result, left[i])
+		i++
+	}
+	for j < len(right) {
+		result = append(result, right[j])
+		j++
+	}
+
+	return result
 }
 
 func main() {