This pull request includes significant changes to the CI workflow and improvements to the sorting algorithm in the main application. The most important changes include adding a new CI workflow for Go projects and refactoring the sorting function to use a more efficient merge sort algorithm.

### CI Workflow Enhancements:
* [`.github/workflows/ci.yml`](diffhunk://#diff-b803fcb7f17ed9235f1e5cb1fcd2f5d3b2838429d4368ae4c57ce4436577f03fR1-R126): Added a new workflow configuration for building, testing, and benchmarking Go projects. This includes steps for setting up Go, running tests, and comparing benchmark results between the current branch and the main branch.

### Code Improvements:
* [`app/main.go`](diffhunk://#diff-d4d93790ef54cfaf975e2d3fb22c460929bc1c34cb55f333508d5a9a0d2f0de6L5-R42): Refactored the `Sort` function to use the merge sort algorithm instead of the bubble sort algorithm. This change improves the efficiency of the sorting process.