@luumarodrigues 2024-10-21
## Resultados do Benchmark

```
goos: linux
goarch: amd64
pkg: tdc24
cpu: AMD EPYC 7763 64-Core Processor                
        │    main     │               current               │
        │   sec/op    │   sec/op     vs base                │
Sort      510.6µ ± 1%   165.5µ ± 1%  -67.60% (p=0.000 n=10)
Sort-2    510.5µ ± 1%   162.9µ ± 1%  -68.09% (p=0.000 n=10)
Sort-4    510.6µ ± 0%   164.7µ ± 2%  -67.75% (p=0.000 n=10)
geomean   510.6µ        164.3µ       -67.81%

        │     main     │                 current                  │
        │     B/op     │      B/op       vs base                  │
Sort      8.000Ki ± 0%   168.070Ki ± 0%  +2000.88% (p=0.000 n=10)
Sort-2    8.000Ki ± 0%   168.070Ki ± 0%  +2000.88% (p=0.000 n=10)
Sort-4    8.000Ki ± 0%   168.070Ki ± 0%  +2000.88% (p=0.000 n=10)
geomean   8.000Ki          168.1Ki       +2000.88%

        │    main    │                  current                  │
        │ allocs/op  │   allocs/op    vs base                    │
Sort      1.000 ± 0%   3013.000 ± 0%  +301200.00% (p=0.000 n=10)
Sort-2    1.000 ± 0%   3013.000 ± 0%  +301200.00% (p=0.000 n=10)
Sort-4    1.000 ± 0%   3013.000 ± 0%  +301200.00% (p=0.000 n=10)
geomean   1.000          3.013k       +301200.00%
```

## Análise do resultado

- **Desempenho aprimorado**: O tempo de execução dos testes na branch atual (current) foi significativamente reduzido em comparação com a branch main, apresentando uma melhoria de cerca de -67,81% no tempo médio de execução (sec/op).
- **Uso de memória aumentado**: Houve um aumento substancial no consumo de memória (B/op), com a branch atual utilizando aproximadamente 2000,88% mais memória do que a branch main.
- **Alocações aumentadas**: O número de alocações por operação (allocs/op) na branch atual cresceu drasticamente, com um aumento de 301200.00% em relação à branch main.
- **Consistência dos resultados**: Os resultados dos benchmarks foram consistentes entre múltiplas execuções, com variações percentuais mínimas observadas tanto no tempo de execução quanto no uso de memória e alocações.
- **Trade-off entre performance e recursos**: A análise sugere um trade-off evidente entre a redução drástica no tempo de execução e o aumento extremamente elevado no consumo de memória e alocações na branch atual.