@luumarodrigues 2024-10-23
## Resultados do Benchmark

```
goos: linux
goarch: amd64
pkg: tdc24
cpu: AMD EPYC 7763 64-Core Processor                
        │    main     │               current               │
        │   sec/op    │   sec/op     vs base                │
Sort      509.9µ ± 0%   165.5µ ± 2%  -67.54% (p=0.000 n=10)
Sort-2    510.8µ ± 0%   163.2µ ± 0%  -68.05% (p=0.000 n=10)
Sort-4    510.5µ ± 0%   166.4µ ± 2%  -67.40% (p=0.000 n=10)
geomean   510.4µ        165.0µ       -67.67%

        │     main     │                 current                  │
        │     B/op     │      B/op       vs base                  │
Sort      8.000Ki ± 0%   168.070Ki ± 0%  +2000.88% (p=0.000 n=10)
Sort-2    8.000Ki ± 0%   168.070Ki ± 0%  +2000.88% (p=0.000 n=10)
Sort-4    8.000Ki ± 0%   168.070Ki ± 0%  +2000.88% (p=0.000 n=10)
geomean   8.000Ki          168.1Ki       +2000.88%

        │    main    │                  current                  │
        │ allocs/op  │   allocs/op    vs base                    │
Sort      1.000 ± 0%   3013.000 ± 0%  +301200.00% (p=0.000 n=10)
Sort-2    1.000 ± 0%   3013.000 ± 0%  +301200.00% (p=0.000 n=10)
Sort-4    1.000 ± 0%   3013.000 ± 0%  +301200.00% (p=0.000 n=10)
geomean   1.000          3.013k       +301200.00%
```

## Análise do resultado

- **Desempenho**: O tempo de execução (sec/op) dos testes "Sort" nas branches "main" e "current" mostrou uma melhoria significativa na nova branch com uma redução de aproximadamente 67.67%.
- **Uso de Memória**: Houve um aumento substancial no consumo de memória (B/op) na branch "current", com um incremento de 2000.88% em comparação à branch "main".
- **Alocação de Objetos**: A quantidade de alocações por operação (allocs/op) disparou na branch "current", apresentando um aumento de 301200.00%, passando de 1 para mais de 3000 alocações.
- **Consistência**: Todos os testes "Sort", "Sort-2" e "Sort-4" seguem a mesma tendência em todas as métricas analisadas, indicando uma modificação no código que afeta consistentemente as operações de ordenação.
- **Conclusão**: Apesar da considerável melhoria no tempo de execução, a nova implementação na branch "current" introduce um custo muito alto em termos de consumo de memória e número de alocações.